chore: Jestify system tests
